chore: #WB-893, remove gradle from scripts

+++ build.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+MVN_MOD_GROUPID=`grep 'modowner=' gradle.properties | sed 's/modowner=//'`
+MVN_MOD_NAME=`grep 'modname=' gradle.properties | sed 's/modname=//'`
+MVN_MOD_VERSION=`grep 'version=' gradle.properties | sed 's/version=//'`
+
 if [ ! -e node_modules ]
 then
   mkdir node_modules
@@ -28,7 +32,7 @@ fi
 
 
 # OVERRIDES VARS
-OVERRIDE_NAME="cg77"
+OVERRIDE_NAME="default"
 for i in "$@"
 do
 case $i in
@@ -44,9 +48,9 @@ done
 MOD_NAME=`grep "modname=" gradle.properties | sed 's/modname=//g'`
 if [ "$OVERRIDE_NAME" = "default" ];
 then
-  export OVERRIDE_MODNAME="$MOD_NAME"
+  export FINAL_MODNAME="$MOD_NAME"
 else
-  export OVERRIDE_MODNAME="$MOD_NAME-$OVERRIDE_NAME"
+  export FINAL_MODNAME="$MOD_NAME-$OVERRIDE_NAME"
 fi
 
 export OVERRIDE_BUILD="build-css"
@@ -59,34 +63,50 @@ clean () {
   rm -rf dist
   rm -rf build
   rm -rf build-css
+  rm -f yarn.lock
   rm -rf deployment/*
 }
 
-init () {
-  echo "[init] Get branch name from jenkins env..."
+doInit () {
+  echo "[init$1][$OVERRIDE_NAME] Get branch name from jenkins env..."
   BRANCH_NAME=`echo $GIT_BRANCH | sed -e "s|origin/||g"`
   if [ "$BRANCH_NAME" = "" ]; then
-    echo "[init] Get branch name from git..."
+    echo "[init$1][$OVERRIDE_NAME] Get branch name from git..."
     BRANCH_NAME=`git branch | sed -n -e "s/^\* \(.*\)/\1/p"`
   fi
-  docker-compose run -e OVERRIDE_NAME=$OVERRIDE_NAME -e OVERRIDE_MODNAME=$OVERRIDE_MODNAME -e BRANCH_NAME=$BRANCH_NAME -e FRONT_TAG=$FRONT_TAG -e NEXUS_ODE_USERNAME=$NEXUS_ODE_USERNAME -e NEXUS_ODE_PASSWORD=$NEXUS_ODE_PASSWORD --rm -u "$USER_UID:$GROUP_GID" gradle sh -c "gradle generateTemplate"
-  docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm install"
+  
+  echo "[init$1][$OVERRIDE_NAME] Generate deployment file from conf.deployment..."
+  mkdir -p deployment/$FINAL_MODNAME
+  cp conf.deployment deployment/$FINAL_MODNAME/conf.json.template
+  sed -i "s/%MODNAME%/${FINAL_MODNAME}/" deployment/$FINAL_MODNAME/conf.json.template
+  sed -i "s/%VERSION%/${MVN_MOD_VERSION}/" deployment/$FINAL_MODNAME/conf.json.template
+  
+  echo "[init$1][$OVERRIDE_NAME] Generate package.json from package.json.template..."
+  NPM_VERSION_SUFFIX=`date +"%Y%m%d%H%M"`
+  cp package.json.template package.json
+  sed -i "s/%generateVersion%/${NPM_VERSION_SUFFIX}/" package.json
+  
+  if [ "$1" == "Dev" ]
+  then
+    sed -i "s/%entcoreCSSVersion%/file:\/home\/node\/entcore-css-lib/" package.json
+  else
+    sed -i "s/%entcoreCSSVersion%/${BRANCH_NAME}/" package.json
+  fi
+
+  echo "[init$1][$OVERRIDE_NAME] Install yarn dependencies..."
+  docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "yarn install"
+}
+
+init() {
+  doInit
 }
 
 initDev () {
-  echo "[init] Get branch name from jenkins env..."
-  BRANCH_NAME=`echo $GIT_BRANCH | sed -e "s|origin/||g"`
-  if [ "$BRANCH_NAME" = "" ]; then
-    echo "[init] Get branch name from git..."
-    BRANCH_NAME=`git branch | sed -n -e "s/^\* \(.*\)/\1/p"`
-  fi
-  docker-compose run -e OVERRIDE_NAME=$OVERRIDE_NAME -e OVERRIDE_MODNAME=$OVERRIDE_MODNAME -e BRANCH_NAME=$BRANCH_NAME -e FRONT_TAG=$FRONT_TAG -e NEXUS_ODE_USERNAME=$NEXUS_ODE_USERNAME -e NEXUS_ODE_PASSWORD=$NEXUS_ODE_PASSWORD --rm -u "$USER_UID:$GROUP_GID" gradle sh -c "gradle generateTemplateDev"
-  docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm install"
+  doInit "Dev"
 }
 
 build () {
   rm -rf build-css/
-  local extras=$1
   #get skins
   dirs=($(ls -d ./skins/*))
   #create build dir var
@@ -99,9 +119,20 @@ build () {
   cp -R $OVERRIDE_SRC/css/* $SCSS_DIR/
   #build css
   docker-compose run -e SKIN_DIR=$SKIN_DIR -e SCSS_DIR=$SCSS_DIR -e DIST_DIR=$OVERRIDE_DIST --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm run release:prepare"
+  status=$?
+  if [ $status != 0 ];
+  then
+    exit $status
+  fi
+
   for dir in "${dirs[@]}"; do
     tmp=`echo $dir | sed 's/.\/skins\///'`
     docker-compose run -e SKIN_DIR=$SKIN_DIR -e SCSS_DIR=$SCSS_DIR -e DIST_DIR=$OVERRIDE_DIST -e SKIN=$tmp  --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm run sass:build:release"
+    status=$?
+    if [ $status != 0 ];
+    then
+      exit $status
+    fi
   done
   #copy override theme
   echo "Merge assets from theme...."
@@ -115,11 +146,16 @@ build () {
   #override i18n
   echo "Merge i18n from theme and platform..."
   docker-compose run -e OVERRIDE_SRC=$OVERRIDE_SRC -e OVERRIDE_DIST=$OVERRIDE_DIST --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm run override:i18n"
+  status=$?
+  if [ $status != 0 ];
+  then
+    exit $status
+  fi
   echo "End merge"
   #set version
   cp node_modules/entcore-css-lib/dist/version.txt $OVERRIDE_DIST/version.txt
   VERSION=`grep "version="  gradle.properties| sed 's/version=//g'`
-  echo "$OVERRIDE_MODNAME=$VERSION `date +'%d/%m/%Y %H:%M:%S'`" >> $OVERRIDE_DIST/version.txt
+  echo "$FINAL_MODNAME=$VERSION `date +'%d/%m/%Y %H:%M:%S'`" >> $OVERRIDE_DIST/version.txt
 }
 
 watch () {
@@ -139,32 +175,33 @@ publishNPM () {
   
   if [ "$OVERRIDE_NAME" != "default" ];
   then
-    # rename npm package name to OVERRIDE_MODNAME in package.json
-    mv package.json package.json.orig
-    sed "0,/theme-open-ent/{s|theme-open-ent|$OVERRIDE_MODNAME|}" package.json.orig > package.json
+    # rename npm package name to FINAL_MODNAME in package.json
+    sed -i "0,/theme-open-ent/{s|theme-open-ent|$FINAL_MODNAME|}" package.json
   fi
 
   docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm publish --tag $LOCAL_BRANCH"
-
-  if [ "$OVERRIDE_NAME" != "default" ];
+  status=$?
+  if [ $status != 0 ];
   then
-    mv package.json.orig package.json
+    exit $status
   fi
+}
+
+archive() {
+  echo "[archive][$OVERRIDE_NAME] Archiving dist folder and conf.j2 file..."
+  tar cfzh ${FINAL_MODNAME}.tar.gz dist/* conf.j2
 }
 
 publishNexus () {
-  if [ -e "?/.gradle" ] && [ ! -e "?/.gradle/gradle.properties" ]
-  then
-    echo "odeUsername=$NEXUS_ODE_USERNAME" > "?/.gradle/gradle.properties"
-    echo "odePassword=$NEXUS_ODE_PASSWORD" >> "?/.gradle/gradle.properties"
-    echo "sonatypeUsername=$NEXUS_SONATYPE_USERNAME" >> "?/.gradle/gradle.properties"
-    echo "sonatypePassword=$NEXUS_SONATYPE_PASSWORD" >> "?/.gradle/gradle.properties"
-  fi
-  docker-compose run -e OVERRIDE_NAME=$OVERRIDE_NAME -e OVERRIDE_MODNAME=$OVERRIDE_MODNAME --rm -u "$USER_UID:$GROUP_GID" gradle sh -c "gradle deploymentJar fatJar publish"
+  case "$MVN_MOD_VERSION" in
+    *SNAPSHOT) nexusRepository='snapshots' ;;
+    *)         nexusRepository='releases' ;;
+  esac
+  mvn deploy:deploy-file --batch-mode -DgroupId=$MVN_MOD_GROUPID -DartifactId=$FINAL_MODNAME -Dversion=$MVN_MOD_VERSION -Dpackaging=tar.gz -Dfile=${FINAL_MODNAME}.tar.gz -DrepositoryId=wse -Durl=https://maven.opendigitaleducation.com/nexus/content/repositories/$nexusRepository/
 }
 
 publishMavenLocal (){
-  docker-compose run -e OVERRIDE_NAME=$OVERRIDE_NAME -e OVERRIDE_MODNAME=$OVERRIDE_MODNAME --rm -u "$USER_UID:$GROUP_GID" gradle sh -c "gradle deploymentJar fatJar publishToMavenLocal"
+  mvn install:install-file --batch-mode -DgroupId=$MVN_MOD_GROUPID -DartifactId=$FINAL_MODNAME -Dversion=$MVN_MOD_VERSION -Dpackaging=tar.gz -Dfile=${FINAL_MODNAME}.tar.gz
 }
 
 for param in "$@"
@@ -187,7 +224,7 @@ do
       rm -rf build-css && build && cp -r dist/* ../recette/assets/themes/theme-open-ent
       ;;
     install)
-      build && publishMavenLocal
+      build && archive && publishMavenLocal
       ;;
     watch)
       watch
@@ -197,6 +234,9 @@ do
       ;;
     lint-fix)
       lint-fix
+      ;;
+    archive)
+      archive
       ;;
     publishNPM)
       publishNPM
